requirements_edu-r.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
braintree==4.28.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==0.61.1
django-braces==1.15.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-embed-video==1.4.9
django-memcache-status==2.3
djangorestframework==3.15.1
flower==2.0.1
humanize==4.9.0
idna==3.6
kombu==5.3.6
oauthlib==3.2.2
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
PyJWT==2.8.0
pylibmc==1.6.3
pymemcache==4.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-memcached==1.62
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uwsgi==2.0.29
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
